+++ src/main/java/org/mae/twg/backend/dto/travel/response/TourDTO.java
@@ -1,12 +1,10 @@
-package org.mae.twg.backend.dto.travel;
+package org.mae.twg.backend.dto.travel.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import org.mae.twg.backend.dto.ModelDTO;
-import org.mae.twg.backend.dto.travel.response.CountryDTO;
-import org.mae.twg.backend.dto.travel.response.HotelLightDTO;
-import org.mae.twg.backend.dto.travel.response.SightLightDTO;
-import org.mae.twg.backend.dto.travel.response.TagDTO;
+import org.mae.twg.backend.dto.travel.response.lightDTOs.HotelLightDTO;
+import org.mae.twg.backend.dto.travel.response.lightDTOs.SightLightDTO;
 import org.mae.twg.backend.exceptions.ObjectNotFoundException;
 import org.mae.twg.backend.models.travel.Tour;
 import org.mae.twg.backend.models.travel.enums.Localization;
@@ -27,6 +25,7 @@ public class TourDTO implements ModelDTO {
     private String introduction;
     private String description;
     private String additional;
+    private Double grade;
     private CountryDTO country;
     private Localization localization;
     private List<String> medias;
